fix(practice): extract logic out of tap and key listener
